feat: move send mail modal
